Moved registrations to their own file to allow handling to be more closer with the actual implementation and registration about what we want to achieve
